examples/python/hello/requirements.txt
strawberry-graphql==0.130.3
